requirements.txt
Django==1.9.1
PyJWT==1.4.0
argparse==1.2.1
coverage==4.0.3
django-allauth==0.24.1
django-debug-toolbar==1.4
httplib2==0.9.2
nose==1.3.7
oauth==1.0.1
oauth-python-twitter==1.0
oauth2==1.9.0.post1
oauthlib==1.0.3
python-openid==2.2.5
python-social-auth==0.2.13
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.0
simplejson==3.8.1
six==1.10.0
sqlparse==0.1.18
wsgiref==0.1.2
